+++ stdnet/lib/lua/orm/delete_query.lua
@@ -1,25 +1,4 @@
 --DELETE A QUERY for a model
-function table_slice (values,i1,i2)
-    local res = {}
-    local n = #values
-    -- default values for range
-    i1 = i1 or 1
-    i2 = i2 or n
-    if i2 < 0 then
-        i2 = n + i2 + 1
-    elseif i2 > n then
-        i2 = n
-    end
-    if i1 < 1 or i1 > n then
-        return {}
-    end
-    local k = 1
-    for i = i1,i2 do
-        res[k] = values[i]
-        k = k + 1
-    end
-    return res
-end
 
 -- Add or remove indices for an instance
 function update_indices(s, bk, id, idkey, indices, uniques)
@@ -43,11 +22,11 @@ rkey = KEYS[2] -- the key where to store the structure containing the resuls
 local s = ARGV[1]
 local length_indices = ARGV[2]
 local idx1 = 2
-local indices = table_slice(ARGV,idx1+1,idx1+length_indices)
-local uniques = table_slice(ARGV,idx1+length_indices+1,idx1+2*length_indices)
+local indices = tabletools.slice(ARGV,idx1+1,idx1+length_indices)
+local uniques = tabletools.slice(ARGV,idx1+length_indices+1,idx1+2*length_indices)
 idx1 = idx1 + 2*length_indices + 1
 local length_multifields = ARGV[idx1]
-local multifields = table_slice(ARGV,idx1+1,idx1+length_multifields)
+local multifields = tabletools.slice(ARGV,idx1+1,idx1+length_multifields)
 i = idx1 + length_multifields
 local idset = bk .. ':id'
 local ty = redis.call('type',rkey)['ok']