+++ stdnet/lib/redis/lua/datefilter.lua
@@ -1,1 +1,0 @@
--- Script for filtering on timestamps